requirements/requirements_few_shot.txt
# Requirements for Few-Shot Learning with CLIP

# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0

# CLIP dependencies
ftfy>=6.1.0
regex>=2022.0.0
tqdm>=4.64.0

# CLIP model (install from GitHub)
git+https://github.com/openai/CLIP.git

# Image processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Numerical computing
numpy>=1.21.0